--- src/main/kotlin/com/example/blinddateserver/application/controller/TestController.kt
@@ -1,16 +1,0 @@
-package com.example.blinddateserver.application.controller
-
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class TestController {
-    @GetMapping("/hello")
-    fun 임시테스트():String{
-        return "혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다." +
-                "혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다."+
-                "혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다."+
-                "혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다."+
-                "혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다.혜진이는 조금 귀엽다."
-    }
-}
