+++ src/test/kotlin/com/example/fleetingserver/FleetingServerApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.example.blinddateserver
+package com.example.fleetingserver
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
